+++ tests/Unit/Tecan/CustomCommand/MllReagentDistributionTest.php
@@ -35,7 +35,7 @@ final class MllReagentDistributionTest extends TestCase
 
         self::assertSame(
             "R;{$sourceRack->name()};;{$sourceRack->type()};{$sourceStartPosition};{$sourceStartPosition};{$targetRack->name()};;{$targetRack->type()};48;73;{$dispenseVolume};{$liquidClass->name()};{$numberOfDitiReuses};{$numberOfMultiDisp};0;50;51;52;53;54;55;56;57;58;59;60;61;62;63;64;65;66;67;68;69;70;71",
-            $command->formatToString()
+            $command->toString()
         );
     }
 }